@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 		<LangVersion>preview</LangVersion>
@@ -15,15 +15,18 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+
+		<PackageReference Include="CoordinateSharp" Version="2.15.2.1" />
+
 		<PackageReference Include="Jint" Version="3.0.0-beta-2041" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="6.0.7" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="6.0.7" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.7" />
-		<PackageReference Include="Newtonsoft.Json" Version="13.0.2-beta1" />
-		<PackageReference Include="NLog.Extensions.Logging" Version="5.0.1" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Proxies" Version="7.0.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.0" />
+		<PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.0" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
+		<PackageReference Include="NLog.Extensions.Logging" Version="5.2.0" />
 		<PackageReference Include="NonSucking.Framework.Extension" Version="0.1.0.74-alpha" />
 		<PackageReference Include="NonSucking.Framework.Serialization" Version="0.1.0.66-alpha" />
-		<PackageReference Include="Quartz" Version="3.3.3" />
+		<PackageReference Include="Quartz" Version="3.5.0" />
 		<PackageReference Include="System.Runtime.Experimental" Version="6.0.2" />
 		<PackageReference Include="NiL.JS" Version="2.5.1600" />
 	</ItemGroup>
