@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <Nullable>enable</Nullable>
 
     <IsPackable>false</IsPackable>
@@ -16,14 +16,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-	  <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.2.0" />
+	  <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
 	  <PackageReference Include="NUnit" Version="3.13.3" />
-	  <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
-	  <PackageReference Include="coverlet.collector" Version="3.1.2">
+	  <PackageReference Include="NUnit3TestAdapter" Version="4.3.1" />
+	  <PackageReference Include="coverlet.collector" Version="3.2.0">
 	    <PrivateAssets>all</PrivateAssets>
 	    <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 	  </PackageReference>
-	  <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.2.0" />
+	  <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.4.0" />
   </ItemGroup>
 
   <ItemGroup>
