@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
 	  <EnablePreviewFeatures>True</EnablePreviewFeatures>
 	  <Nullable>enable</Nullable>
 	  <!--<IncludePackageReferencesDuringMarkupCompilation>true</IncludePackageReferencesDuringMarkupCompilation>
@@ -20,9 +20,9 @@
   <ItemGroup>
     <PackageReference Include="Azura" Version="0.9.0" />
     <PackageReference Include="Azura.Generator" Version="0.9.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.2-beta1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
     <PackageReference Include="NonSucking.Framework.Serialization" Version="0.1.0.66-alpha" />
-    <PackageReference Include="NLog" Version="5.0.1" />
+    <PackageReference Include="NLog" Version="5.1.0" />
   </ItemGroup>
 
   <ItemGroup>
