@@ -7,6 +7,10 @@ using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
+using NLog;
+
+using System.Runtime.InteropServices;
+using System.Xml.Linq;
 
 namespace AppBrokerASP.State;
 
@@ -15,18 +19,38 @@ public class DeviceStateManager : IDeviceStateManager
     public event EventHandler<StateChangeArgs>? StateChanged;
 
     private readonly ConcurrentDictionary<long, Dictionary<string, JToken>> deviceStates = new();
+    private readonly Dictionary<string, Dictionary<string, HashSet<string>>> propertyNameMappings = new();
+    private readonly FileSystemWatcher fileSystemWatcher;
+    private readonly Logger logger;
 
     public bool ManagesDevice(long id) => deviceStates.ContainsKey(id);
 
     public DeviceStateManager()
     {
+        logger = LogManager.GetCurrentClassLogger();
         using var ctx = DbProvider.BrokerDbContext;
-        foreach (var item in ctx.Devices.Where(x=> !string.IsNullOrWhiteSpace(x.LastState)))
+        foreach (var item in ctx.Devices.Where(x => !string.IsNullOrWhiteSpace(x.LastState)))
         {
             deviceStates[item.Id] = JsonConvert.DeserializeObject<Dictionary<string, JToken>>(item.LastState!)!;
         }
-        
+
+        Directory.CreateDirectory("PropertyMappings");
+        fileSystemWatcher = new FileSystemWatcher(new DirectoryInfo("PropertyMappings").FullName, "*.json")
+        {
+            NotifyFilter = NotifyFilters.FileName |
+                           NotifyFilters.LastWrite |
+                           NotifyFilters.Security
+        };
+        fileSystemWatcher.Changed += FileChanged;
+        fileSystemWatcher.Created += FileChanged;
+        fileSystemWatcher.EnableRaisingEvents = true;
+        foreach (var path in Directory.EnumerateFiles("PropertyMappings", "*.json", SearchOption.AllDirectories))
+        {
+            ReadPropertyMappings(path);
+        }
+
     }
+
 
     //TODO Router or Mainspower get from Zigbee2Mqtt
     public Dictionary<string, JToken>? GetCurrentState(long id)
@@ -65,6 +89,7 @@ public class DeviceStateManager : IDeviceStateManager
 
     public void SetSingleState(long id, string propertyName, JToken newVal)
     {
+        propertyName = MapValueName(id, propertyName);
 
         if (deviceStates.TryGetValue(id, out var oldState))
         {
@@ -91,9 +116,34 @@ public class DeviceStateManager : IDeviceStateManager
             device.SendDataToAllSubscribers();
     }
 
+    public void PushNewState(long id, string propertyName, JToken newVal)
+    {
+        propertyName = MapValueName(id, propertyName);
+
+        if (deviceStates.TryGetValue(id, out var oldState))
+        {
+            if (oldState.ContainsKey(propertyName) && JToken.DeepEquals(oldState[propertyName], newVal))
+                return;
+
+            SetNewState(id, propertyName, newVal, oldState);
+        }
+        else
+        {
+            InstanceContainer.Instance.HistoryManager.StoreNewState(id, propertyName, null, newVal);
+            AddStatesForBackwartsCompatibilityForOldApp(id, propertyName, newVal);
+
+            deviceStates[id][propertyName] = newVal;
+        }
+
+        if (InstanceContainer.Instance.DeviceManager.Devices.TryGetValue(id, out var device))
+            device.SendDataToAllSubscribers();
+
+        StoreLastState(id, deviceStates[id], device);
+    }
 
     public void PushNewState(long id, Dictionary<string, JToken> newState)
     {
+        newState = newState.ToDictionary(x => MapValueName(id, x.Key), x => x.Value);
         if (deviceStates.TryGetValue(id, out var oldState))
         {
             List<string> changedKeys = new();
@@ -125,12 +175,13 @@ public class DeviceStateManager : IDeviceStateManager
 
         if (InstanceContainer.Instance.DeviceManager.Devices.TryGetValue(id, out var device))
             device.SendDataToAllSubscribers();
-        
+
         StoreLastState(id, deviceStates[id], device);
     }
 
     private void SetNewState(long id, string propertyName, JToken newVal, Dictionary<string, JToken> state)
     {
+        propertyName = MapValueName(id, propertyName);
         var oldVal = state.GetValueOrDefault(propertyName);
         IInstanceContainer.Instance.HistoryManager.StoreNewState(id, propertyName, oldVal, newVal);
         StateChanged?.Invoke(this, new(id, propertyName, oldVal, newVal));
@@ -172,5 +223,53 @@ public class DeviceStateManager : IDeviceStateManager
         if (string.IsNullOrWhiteSpace(newPropName))
             return;
         SetSingleState(id, newPropName, value);
+    }
+
+
+
+    private void FileChanged(object sender, FileSystemEventArgs e)
+    {
+        propertyNameMappings.Clear();
+        foreach (var path in Directory.EnumerateFiles("PropertyMappings", "*.json", SearchOption.AllDirectories))
+            ReadPropertyMappings(path);
+    }
+
+    private void ReadPropertyMappings(string path)
+    {
+        using var str = File.OpenRead(path);
+        var res = System.Text.Json.JsonSerializer.Deserialize<Dictionary<string, Dictionary<string, HashSet<string>>>>(str);
+
+        foreach (var (deviceType, mappings) in res)
+        {
+            ref var mapsForType = ref CollectionsMarshal.GetValueRefOrAddDefault(propertyNameMappings, deviceType, out _);
+            mapsForType ??= new();
+
+            logger.Trace($"Add Mapping for {deviceType}({mappings.Count})");
+            foreach (var (propertyName, oldNames) in mappings)
+            {
+                mapsForType[propertyName] = oldNames;
+            }
+        }
+    }
+
+    public string MapValueName(long id, string name)
+    {
+        if (IInstanceContainer.Instance.DeviceManager.Devices.TryGetValue(id, out var device))
+        {
+            foreach (var deviceType in device.TypeNames)
+            {
+                if (propertyNameMappings.TryGetValue(deviceType, out var mapsForType))
+                {
+                    foreach (var (newPropName, oldValueNames) in mapsForType)
+                    {
+                        if (oldValueNames.Contains(name))
+                        {
+                            return newPropName;
+                        }
+                    }
+                }
+            }
+        }
+        return name;
     }
 }
