@@ -8,17 +8,18 @@ using Newtonsoft.Json.Linq;
 using Microsoft.AspNetCore.Server.IISIntegration;
 using System.Globalization;
 using AppBroker.Core;
+using AppBroker.Core.Database;
 
 namespace AppBrokerASP.Zigbee2Mqtt;
 
-public class Zigbee2MqttManager : IAsyncDisposable
+public class Zigbee2MqttManager : IAsyncDisposable, IZigbee2MqttManager
 {
-    private IManagedMqttClient? mqtt;
+    public IManagedMqttClient? MQTTClient { get; set; }
     Device[]? devices = null;
     private readonly Zigbee2MqttConfig config;
     private readonly Logger logger;
     private readonly Dictionary<string, long> friendlyNameToIdMapping = new();
-    private readonly Stack<(string, string)> cachedBeforeConnect = new();
+    private readonly Stack<(string, string, string)> cachedBeforeConnect = new();
 
     public Zigbee2MqttManager(Zigbee2MqttConfig zigbee2MqttConfig)
     {
@@ -28,12 +29,12 @@ public class Zigbee2MqttManager : IAsyncDisposable
 
     public async Task Subscribe()
     {
-        if (mqtt is null)
+        if (MQTTClient is null)
             return;
 
         try
         {
-            await mqtt.SubscribeAsync("zigbee2mqtt/#");
+            await MQTTClient.SubscribeAsync("zigbee2mqtt/#");
 
         }
         catch (Exception ex)
@@ -44,8 +45,8 @@ public class Zigbee2MqttManager : IAsyncDisposable
 
     public async Task<IManagedMqttClient> Connect()
     {
-        if (mqtt is not null)
-            return mqtt;
+        if (MQTTClient is not null)
+            return MQTTClient;
 
         var mqttFactory = new MqttFactory();
         var managedMqttClient = mqttFactory.CreateManagedMqttClient();
@@ -58,57 +59,87 @@ public class Zigbee2MqttManager : IAsyncDisposable
             .Build();
 
         await managedMqttClient.StartAsync(managedMqttClientOptions);
-        mqtt = managedMqttClient;
+        MQTTClient = managedMqttClient;
 
-        mqtt.ApplicationMessageReceivedAsync += Mqtt_ApplicationMessageReceivedAsync;
+        MQTTClient.ApplicationMessageReceivedAsync += Mqtt_ApplicationMessageReceivedAsync;
 
-        return mqtt;
+        return MQTTClient;
     }
 
-    private Task Mqtt_ApplicationMessageReceivedAsync(MqttApplicationMessageReceivedEventArgs e)
+    private async Task Mqtt_ApplicationMessageReceivedAsync(MqttApplicationMessageReceivedEventArgs e)
     {
         var topic = e.ApplicationMessage.Topic;
+        if (!topic.StartsWith("zigbee2mqtt/"))
+            return;
+
         var payload = e.ApplicationMessage.ConvertPayloadToString();
 
-        switch (topic)
+        var splitted = topic.Split('/', 3);
+
+        var deviceName = splitted[1];
+        if (splitted.Length < 3)
         {
-            case "zigbee2mqtt/bridge/devices":
+            Console.WriteLine($"[{topic}] {payload}");
+            if (devices is null)
+            {
+                await MQTTClient.SubscribeAsync("zigbee2mqtt/bridge/devices");
+                logger.Trace($"Got state before device {deviceName}");
+                cachedBeforeConnect.Push((topic, deviceName, payload));
+                return;
+            }
+            TryInterpretTopicAsStateUpdate(topic, deviceName, payload);
+            return;
+
+        }
+
+        var method = splitted[2];
+        switch (method)
+        {
+            case "devices":
                 devices = JsonConvert.DeserializeObject<Device[]>(payload);
-                foreach (var item in devices!)
+                using (var ctx = DbProvider.BrokerDbContext)
                 {
-                    var id = long.Parse(item.IEEEAddress[2..], NumberStyles.HexNumber);
-                    if (item.Type == DeviceType.Coordinator || IInstanceContainer.Instance.DeviceManager.Devices.ContainsKey(id))
-                        continue;
-                    var dev = new Zigbee2MqttDevice(item, id, mqtt!);
-                    InstanceContainer.Instance.DeviceManager.AddNewDevice(dev);
-                    friendlyNameToIdMapping[item.FriendlyName] = dev.Id;
+
+                    foreach (var item in devices!)
+                    {
+                        var id = long.Parse(item.IEEEAddress[2..], NumberStyles.HexNumber);
+                        var dbDevice = ctx.Devices.FirstOrDefault(x => x.Id == id);
+                        if (dbDevice is not null && !string.IsNullOrWhiteSpace(dbDevice.FriendlyName) && dbDevice.FriendlyName != item.FriendlyName)
+                        {
+                            await MQTTClient.EnqueueAsync("zigbee2mqtt/bridge/request/device/rename", $"{{\"from\": \"{item.IEEEAddress}\", \"to\": \"{dbDevice.FriendlyName}\"}}");
+                            item.FriendlyName = dbDevice.FriendlyName;
+                        }
+
+                        friendlyNameToIdMapping[item.FriendlyName] = id;
+                        if (item.Type == DeviceType.Coordinator || IInstanceContainer.Instance.DeviceManager.Devices.ContainsKey(id))
+                            continue;
+                        var dev = new Zigbee2MqttDevice(item, id);
+                        InstanceContainer.Instance.DeviceManager.AddNewDevice(dev);
+                    }
                 }
                 while (cachedBeforeConnect.TryPop(out var item))
-                    TryInterpretTopicAsStateUpdate(item.Item1, item.Item2);
+                    TryInterpretTopicAsStateUpdate(item.Item1, item.Item2, item.Item3);
 
                 break;
 
-            case "zigbee2mqtt/bridge/config":
+            case "config":
                 var b = JsonConvert.DeserializeObject<BridgeConfig>(payload);
                 ;
                 break;
 
-            case "zigbee2mqtt/bridge/info":
+            case "info":
                 var c = JsonConvert.DeserializeObject<BridgeInfo>(payload);
                 ;
                 break;
 
-            case "zigbee2mqtt/bridge/groups":
+            case "groups":
                 var d = JsonConvert.DeserializeObject<Group[]>(payload);
-                ;
                 break;
-
-            case "zigbee2mqtt/bridge/state":
-                _ = Enum.TryParse<BridgeState>(payload, out var f);
-                ;
-                break;
-
-            case "zigbee2mqtt/bridge/logging":
+            //case "state":
+            //    _ = Enum.TryParse<AvailabilityState>(payload, out var f);
+            //    ;
+            //    break;
+            case "logging":
                 var g = JsonConvert.DeserializeObject<LogMessage>(payload);
                 var lvl = g!.Level switch
                 {
@@ -120,29 +151,31 @@ public class Zigbee2MqttManager : IAsyncDisposable
                 _ = logger.ForLogEvent(lvl).Message(g.Message).Callsite();
 
                 break;
-
-            case "zigbee2mqtt/bridge/extensions":
-            default:
-                Console.WriteLine($"[{topic}] {payload}");
-                if (devices is null)
+            //"zigbee2mqtt/MY_DEVICE/availability"
+            case "availability":
+                if (Enum.TryParse<AvailabilityState>(payload, out var available)
+                    && friendlyNameToIdMapping.TryGetValue(deviceName, out var deviceId))
                 {
-                    cachedBeforeConnect.Push((topic, payload));
-                    break;
+                    InstanceContainer
+                        .Instance
+                        .DeviceStateManager
+                        .PushNewState(deviceId, "available", available == AvailabilityState.Online);
                 }
-                TryInterpretTopicAsStateUpdate(topic, payload);
+                break;
+            case "extensions":
+            default:
                 break;
         }
 
-        return Task.CompletedTask;
+        return;
     }
 
-    private void TryInterpretTopicAsStateUpdate(string topic, string payload)
+    private void TryInterpretTopicAsStateUpdate(string topic, string deviceName, string payload)
     {
         var zigbeeLength = "zigbee2mqtt/".Length;
         if (topic.Length > zigbeeLength)
         {
-            var maybeFriendlyName = topic["zigbee2mqtt/".Length..];
-            if (friendlyNameToIdMapping.TryGetValue(maybeFriendlyName, out var id))
+            if (friendlyNameToIdMapping.TryGetValue(deviceName, out var id))
             {
                 InstanceContainer
                     .Instance
@@ -159,9 +192,9 @@ public class Zigbee2MqttManager : IAsyncDisposable
 
         foreach (var item in GetFeatures<BinaryFeature>(zdev.device.Definition.Exposes))
         {
-            if(customStates.TryGetValue(item.Name, out var token))
+            if (customStates.TryGetValue(item.Property, out var token))
                 customStates[item.Name] = item.ConvertToBool(token.ToOObject());
-            
+
         }
         return customStates;
     }
@@ -186,9 +219,18 @@ public class Zigbee2MqttManager : IAsyncDisposable
         }
     }
 
+    public async Task<bool> SetState(string deviceName, string propertyName, JToken newValue)
+    {
+        if (MQTTClient is null)
+            return false;
+
+        await MQTTClient.EnqueueAsync($"zigbee2mqtt/{deviceName}/set/{propertyName}", newValue.ToString());
+        return true;
+    }
+
     public async ValueTask DisposeAsync()
     {
-        var local = mqtt;
+        var local = MQTTClient;
         if (local is not null)
         {
             await local.StopAsync();
